eth/catalyst: move Register to api.go
